# Truncate range tombstones by leveraging InternalKeys