--- a/db/table_cache.cc
+++ b/db/table_cache.cc
@@ -184,8 +184,8 @@ InternalIterator* TableCache::NewIterator(
     RangeDelAggregator* range_del_agg, const SliceTransform* prefix_extractor,
     TableReader** table_reader_ptr, HistogramImpl* file_read_hist,
     bool for_compaction, Arena* arena, bool skip_filters, int level,
-    const Slice* smallest_compaction_key,
-    const Slice* largest_compaction_key) {
+    const InternalKey* smallest_compaction_key,
+    const InternalKey* largest_compaction_key) {
   PERF_TIMER_GUARD(new_table_iterator_nanos);
 
   Status s;
@@ -270,12 +270,11 @@ InternalIterator* TableCache::NewIterator(
       if (s.ok()) {
         const InternalKey* smallest = &file_meta.smallest;
         const InternalKey* largest = &file_meta.largest;
-        InternalKey smallest_compaction_ikey, largest_compaction_ikey;
         if (smallest_compaction_key != nullptr) {
-          smallest_compaction_ikey.Set(*smallest_compaction_key, kMaxSequenceNumber, kTypeRangeDeletion);
+          smallest = smallest_compaction_key;
         }
         if (largest_compaction_key != nullptr) {
-          largest_compaction_ikey.Set(*largest_compaction_key, kMaxSequenceNumber, kTypeRangeDeletion);
+          largest = largest_compaction_key;
         }
         s = range_del_agg->AddTombstones(std::move(range_del_iter), smallest,
                                          largest);