Summary: To more accurately truncate range tombstones at SST boundaries,
we now represent them in RangeDelAggregator using InternalKeys, which
are end-key-exclusive as they were before this change.

During compaction, "atomic compaction unit boundaries" (the range of
keys contained in neighbouring and overlaping SSTs) are propagated down
to RangeDelAggregator to truncate range tombstones at those boundariies
instead. See https://github.com/facebook/rocksdb/pull/4432#discussion_r221072219 and https://github.com/facebook/rocksdb/pull/4432#discussion_r221138683
for motivating examples.

Test Plan: make -j64 check ; make -j64 range_del_aggregator_bench && ./range_del_aggregator_bench (before and after PR)

Before:
```
=========================
Results: 
=========================
AddTombstones:           578.669 us
ShouldDelete (first):    0.32449 us
```

After:
```
=========================
Results:
=========================
AddTombstones:           649.471 us
ShouldDelete (first):    0.29454 us
```

So using `InternalKeyComparator` results in a slight perf regression.