@ajkr 2018-09-27
also let's verify no perf regression compared to before this PR, and include results in "Test Plan". Here's a few commands I've used before:

- build db_bench:
```
$ DEBUG_LEVEL=0 make -j64 db_bench
```

- benchmark setup:
```
$ ./db_bench -db=/dev/shm/5k-rts -use_existing_db=false -benchmarks=filluniquerandom -write_buffer_size=1048576 -compression_type=lz4 -target_file_size_base=1048576 -max_bytes_for_level_base=4194304 -value_size=112 -key_size=16 -block_size=4096 -level_compaction_dynamic_level_bytes=true -num=5000000 -max_background_jobs=12 -benchmark_write_rate_limit=20971520 -range_tombstone_width=100 -writes_per_range_tombstone=100 -max_num_range_tombstones=50000 -bloom_bits=8
```

- benchmark random reads:

```
$ ./db_bench -db=/dev/shm/5k-rts/ -use_existing_db=true -benchmarks=readrandom -disable_auto_compactions=true -num=5000000 -reads=100000 -threads=32
```

- benchmark iteration:

```
$ ./db_bench -db=/dev/shm/5k-rts/ -use_existing_db=true -benchmarks=seekrandom -seek_nexts=10 -disable_auto_compactions=true -num=5000000 -reads=100000 -threads=32
```