@abhimadan 2018-10-01
I've revamped this PR after all the discussion above to avoid extraneous string copying, and to be more generous with L1+ tombstone truncation during compactions. I want to write tests for the atomic compaction unit boundary logic, but at least RangeDelAggregator should have more comprehensive tests now.